#5 DeleteDelivery atualizado para verificação de excepção

+++ WarehouseManagement/Domain/Deliveries/DeliveryService.cs
@@ -60,6 +60,10 @@ namespace EletricGo.Domain.Deliveries
         public async Task<DeliveryDTO> DeleteDelivery(DeliveryDTO deliveryDTO)
         {
             var delivery = await _deliveryRepository.GetByID(new DeliveryID(deliveryDTO.deliveryID));
+
+            if (delivery == null)
+                return null;
+            
             _deliveryRepository.Delete(delivery);
             await this._unitOfWork.CommitAsync();
             return delivery.toDeliveryDTO();
